Make explorer usable with Kopano Rest and Konnect
This changeset modifies URLs to be able to use Kopano REST API and
Konnect as authentication source.

For now the unauthenticated mode is not supported (requires a proxy
service which we do not have yet).

All in all, this was quite simple with only minor compatibility fixes in
the Graph Explorer implementation while the Graph API itself is
untouched.